Refactor runner_shared logging helpers
