+++ projects/04-llm-adapter-shadow/src/llm_adapter/runner_shared/__init__.py
@@ -6,14 +6,9 @@ import threading
 import time
 
 from .costs import estimate_cost, provider_model
-from .logging import (
-    error_family,
-    log_provider_call,
-    log_provider_skipped,
-    log_run_metric,
-    MetricsPath,
-    resolve_event_logger,
-)
+from .logging.base import MetricsPath, resolve_event_logger
+from .logging.events import log_provider_call, log_provider_skipped, log_run_metric
+from .logging.status import error_family
 from .rate_limiter import RateLimiter, resolve_rate_limiter
 
 __all__ = [
