+++ projects/04-llm-adapter-shadow/src/llm_adapter/runner_shared/logging/__init__.py
@@ -1,0 +1,15 @@
+"""Structured logging helpers for runner modules."""
+from __future__ import annotations
+
+from .base import MetricsPath, resolve_event_logger
+from .events import log_provider_call, log_provider_skipped, log_run_metric
+from .status import error_family
+
+__all__ = [
+    "MetricsPath",
+    "resolve_event_logger",
+    "error_family",
+    "log_provider_skipped",
+    "log_provider_call",
+    "log_run_metric",
+]
