+++ projects/04-llm-adapter-shadow/tests/sync_metrics/test_errors.py
@@ -9,7 +9,7 @@ from src.llm_adapter.errors import AllFailedError, FatalError, TimeoutError
 from src.llm_adapter.parallel_exec import ParallelExecutionError
 from src.llm_adapter.provider_spi import ProviderRequest, ProviderResponse, ProviderSPI
 from src.llm_adapter.runner_config import RunnerMode
-from src.llm_adapter.runner_shared.logging import error_family
+from src.llm_adapter.runner_shared.logging.status import error_family
 
 from ..parallel_helpers import _StaticProvider
 from .helpers import (
